refactor: 채팅방 멤버 명령 책임 분리
- ChatService에 남아 있던 방 나가기와 멤버 강퇴 명령 흐름을 ChatRoomMemberCommandService로 이동
- direct 나가기는 leftAt 상태 변경으로 유지하고 group 나가기는 멤버 삭제 정책을 그대로 둠
- club group 나가기 거부, 비그룹 강퇴 거부, 자기 자신과 방장 강퇴 거부 정책을 기존 응답 코드와 함께 보존
