+++ src/main/java/gg/agit/konect/domain/chat/service/ChatService.java
@@ -80,6 +80,7 @@ public class ChatService {
     private final UserRepository userRepository;
     private final ChatPresenceService chatPresenceService;
     private final ChatRoomMembershipService chatRoomMembershipService;
+    private final ChatRoomMemberCommandService chatRoomMemberCommandService;
     private final ChatRoomSummaryService chatRoomSummaryService;
     private final ChatSearchService chatSearchService;
     private final ChatMessagePageResolver chatMessagePageResolver;
@@ -177,37 +178,12 @@ public class ChatService {
 
     @Transactional
     public void leaveChatRoom(Integer userId, Integer roomId) {
-        ChatRoom room = chatRoomRepository.findById(roomId)
-            .orElseThrow(() -> CustomException.of(NOT_FOUND_CHAT_ROOM));
-
-        if (room.isClubGroupRoom()) {
-            throw CustomException.of(CANNOT_LEAVE_GROUP_CHAT_ROOM);
-        }
-
-        ChatRoomMember member = getRoomMember(roomId, userId);
-        if (room.isDirectRoom()) {
-            member.leaveDirectRoom(LocalDateTime.now());
-            return;
-        }
-
-        chatRoomMemberRepository.deleteByChatRoomIdAndUserId(roomId, userId);
+        chatRoomMemberCommandService.leaveChatRoom(userId, roomId);
     }
 
     @Transactional
     public void kickMember(Integer requesterId, Integer roomId, Integer targetUserId) {
-        ChatRoom room = chatRoomRepository.findById(roomId)
-            .orElseThrow(() -> CustomException.of(NOT_FOUND_CHAT_ROOM));
-
-        validateGroupRoomForKick(room);
-        validateNotSelfKick(requesterId, targetUserId);
-
-        ChatRoomMember requester = getRoomMember(roomId, requesterId);
-        validateKickAuthority(requester);
-
-        ChatRoomMember target = getRoomMember(roomId, targetUserId);
-        validateNotOwnerTarget(target);
-
-        chatRoomMemberRepository.deleteByChatRoomIdAndUserId(roomId, targetUserId);
+        chatRoomMemberCommandService.kickMember(requesterId, roomId, targetUserId);
     }
 
     public ChatRoomsSummaryResponse getChatRooms(Integer userId) {
@@ -1321,30 +1297,6 @@ public class ChatService {
             throw CustomException.of(FORBIDDEN_CHAT_ROOM_ACCESS);
         }
         return partner;
-    }
-
-    private void validateGroupRoomForKick(ChatRoom room) {
-        if (!room.isGroupRoom() || room.isClubGroupRoom()) {
-            throw CustomException.of(CANNOT_KICK_IN_NON_GROUP_ROOM);
-        }
-    }
-
-    private void validateNotSelfKick(Integer requesterId, Integer targetUserId) {
-        if (requesterId.equals(targetUserId)) {
-            throw CustomException.of(CANNOT_KICK_SELF);
-        }
-    }
-
-    private void validateKickAuthority(ChatRoomMember requester) {
-        if (!requester.isOwner()) {
-            throw CustomException.of(FORBIDDEN_CHAT_ROOM_KICK);
-        }
-    }
-
-    private void validateNotOwnerTarget(ChatRoomMember target) {
-        if (target.isOwner()) {
-            throw CustomException.of(CANNOT_KICK_ROOM_OWNER);
-        }
     }
 
     private void recordPresenceSafely(Integer roomId, Integer userId) {
